Improve the seeding with checking if astronaut has another mission, and if so add

+++ client/src/components/Mission.js
@@ -28,47 +28,49 @@ const Mission = ({ mission, astronauts, setAstronauts, onUpdateMissions, onRemov
     fetch(`/missions/${id}`, {
       method: 'DELETE'
     })
-    .then(res => res.json())
-    .then(() => {
-      onRemoveMission(mission);
+    .then(res => {
+      if (res.ok) {
+        onRemoveMission(mission);
 
-      // Remove astronauts associated with the mission
-      if (crew.length > 0) {
-        const updatedAstronauts = [...astronauts]; // Create a copy of the astronauts list
+        // Remove astronauts associated with the mission
+        if (crew.length > 0) {
+          const updatedAstronauts = [...astronauts]; // Create a copy of the astronauts list
 
-        crew.forEach(astroName => {
-          const astronaut = updatedAstronauts.find(a => a.name === astroName);
-          if (astronaut) {
-            if (astronaut.missions.length === 1) {
-              // Delete astronaut if this is their only mission
-              fetch(`/astronauts/${astronaut.id}`, {
-                method: 'DELETE'
-              })
-              .then(res => res.json())
-              .then(() => {
-                const index = updatedAstronauts.findIndex(a => a.id === astronaut.id);
-                if (index > -1) updatedAstronauts.splice(index, 1); // Remove astronaut from the list
-                setAstronauts(updatedAstronauts); // Update state
-              });
-            } else {
-              // Update astronaut's missions if they have other missions
-              const updatedMissions = astronaut.missions.filter(m => m !== name);
-              fetch(`/astronauts/${astronaut.id}`, {
-                method: 'PATCH',
-                headers: {
-                  'Content-Type': 'application/json'
-                },
-                body: JSON.stringify({ missions: updatedMissions })
-              })
-              .then(res => res.json())
-              .then(updatedAstronaut => {
-                const index = updatedAstronauts.findIndex(a => a.id === updatedAstronaut.id);
-                if (index > -1) updatedAstronauts[index] = updatedAstronaut; // Update astronaut in the list
-                setAstronauts(updatedAstronauts); // Update state
-              });
+          crew.forEach(astroName => {
+            const astronaut = updatedAstronauts.find(a => a.name === astroName);
+            if (astronaut) {
+              if (astronaut.missions.length === 1) {
+                // Delete astronaut if this is their only mission
+                fetch(`/astronauts/${astronaut.id}`, {
+                  method: 'DELETE'
+                })
+                .then(res => {
+                  if (res.ok) {
+                    const index = updatedAstronauts.findIndex(a => a.id === astronaut.id);
+                    if (index > -1) updatedAstronauts.splice(index, 1); // Remove astronaut from the list
+                    setAstronauts(updatedAstronauts); // Update state
+                  }
+                });
+              } else {
+                // Update astronaut's missions if they have other missions
+                const updatedMissions = astronaut.missions.filter(m => m !== name);
+                fetch(`/astronauts/${astronaut.id}`, {
+                  method: 'PATCH',
+                  headers: {
+                    'Content-Type': 'application/json'
+                  },
+                  body: JSON.stringify({ missions: updatedMissions })
+                })
+                .then(res => res.json())
+                .then(updatedAstronaut => {
+                  const index = updatedAstronauts.findIndex(a => a.id === updatedAstronaut.id);
+                  if (index > -1) updatedAstronauts[index] = updatedAstronaut; // Update astronaut in the list
+                  setAstronauts(updatedAstronauts); // Update state
+                });
+              }
             }
-          }
-        });
+          });
+        }
       }
     });
   }
